pyproject.toml
[tool.poetry]
name = "rest-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "rest_api"}]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
pyyaml = "^6.0"
uritemplate = "^4.1.1"
drf-spectacular = "^0.26.2"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
